Setup locale at boot even if user boostrapping is enabled, setup also in login
Fixes a bug where the locale setup function wouldn't dispatch the `setLocale` action
in case when the translations aren't server boostrapped (i.e., `window.i18nLocaleStrings`
is not present) and when user boostrapping is enabled ( see the `! config.isEnabled( 'wpcom-user-bootstrap' )`
condition).

In that case, the `state.ui.language` Redux state would be left in default state, which
in combination with persistence, can lead to weird states. Indexed DB contains some locale,
then I change my locale in another session, the load Calypso again with state info in
persisted state. Things that depend on `state.ui.language` (moment.js localization, TinyMCE
localization, REST API `locale` query arg...) would use the old locale, while `i18n-calypso`
translations use the new locale.

This patch also extracts the `setupLocale` function to a separate module and uses it in the
login entrypoint, too.
