+++ client/boot/common.js
@@ -6,7 +6,7 @@ import debugFactory from 'debug';
 import page from 'page';
 import { parse } from 'qs';
 import url from 'url';
-import { get, startsWith } from 'lodash';
+import { startsWith } from 'lodash';
 import React from 'react';
 import ReactDom from 'react-dom';
 import Modal from 'react-modal';
@@ -42,20 +42,10 @@ import { getHappychatAuth } from 'state/happychat/utils';
 import wasHappychatRecentlyActive from 'state/happychat/selectors/was-happychat-recently-active';
 import { setRoute as setRouteAction } from 'state/ui/actions';
 import { getSelectedSiteId, getSectionName } from 'state/ui/selectors';
-import { setLocale, setLocaleRawData } from 'state/ui/language/actions';
 import { setNextLayoutFocus, activateNextLayoutFocus } from 'state/ui/layout-focus/actions';
 import setupGlobalKeyboardShortcuts from 'lib/keyboard-shortcuts/global';
-import { loadUserUndeployedTranslations } from 'lib/i18n-utils/switch-locale';
 
 const debug = debugFactory( 'calypso' );
-
-const switchUserLocale = ( currentUser, reduxStore ) => {
-	const { localeSlug, localeVariant } = currentUser.get();
-
-	if ( localeSlug ) {
-		reduxStore.dispatch( setLocale( localeSlug, localeVariant ) );
-	}
-};
 
 const setupContextMiddleware = reduxStore => {
 	page( '*', ( context, next ) => {
@@ -171,29 +161,6 @@ const clearNoticesMiddleware = () => {
 const unsavedFormsMiddleware = () => {
 	// warn against navigating from changed, unsaved forms
 	page.exit( '*', checkFormHandler );
-};
-
-export const locales = ( currentUser, reduxStore ) => {
-	debug( 'Executing Calypso locales.' );
-
-	if ( window.i18nLocaleStrings ) {
-		const i18nLocaleStringsObject = JSON.parse( window.i18nLocaleStrings );
-		reduxStore.dispatch( setLocaleRawData( i18nLocaleStringsObject ) );
-		const languageSlug = get( i18nLocaleStringsObject, [ '', 'localeSlug' ] );
-		if ( languageSlug ) {
-			debug( 'Checking for load-user-translations parameter' );
-			loadUserUndeployedTranslations( languageSlug );
-		}
-	}
-
-	// Use current user's locale if it was not bootstrapped (non-ssr pages)
-	if (
-		! window.i18nLocaleStrings &&
-		! config.isEnabled( 'wpcom-user-bootstrap' ) &&
-		currentUser.get()
-	) {
-		switchUserLocale( currentUser, reduxStore );
-	}
 };
 
 export const utils = () => {
