+++ client/boot/locale.js
@@ -1,0 +1,27 @@
+/**
+ * External dependencies
+ */
+import { get } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import { loadUserUndeployedTranslations } from 'lib/i18n-utils/switch-locale';
+import { setLocale, setLocaleRawData } from 'state/ui/language/actions';
+
+export const setupLocale = ( currentUser, reduxStore ) => {
+	if ( window.i18nLocaleStrings ) {
+		// Use the locale translation data that were boostrapped by the server
+		const i18nLocaleStringsObject = JSON.parse( window.i18nLocaleStrings );
+		reduxStore.dispatch( setLocaleRawData( i18nLocaleStringsObject ) );
+		const languageSlug = get( i18nLocaleStringsObject, [ '', 'localeSlug' ] );
+		if ( languageSlug ) {
+			loadUserUndeployedTranslations( languageSlug );
+		}
+	} else if ( currentUser && currentUser.localeSlug ) {
+		// Use the current user's and load traslation data with a fetch request
+		reduxStore.dispatch( setLocale( currentUser.localeSlug, currentUser.localeVariant ) );
+	}
+
+	// If user is logged out and translations are not boostrapped, we assume default locale
+};
